I made the data scraping function that scrapes from JHElectronica
I wrote the code for the data scraping from JHElectronica on server side. The function follows the url address given, gets the body of the page and extracts the elements for title, price and img src. Since there is another scraping function, that scrapes from google, I combine the data in a common array and the array is being sent to the client side of the application. There a filter function is to take place and filter through the array and return an result accordingly to the user input