I tried to make the filtering function user triggered
I tried to make the filtering function to be called by user interaction. I set a state variable that gets the user input. Then I assigned the filtering function to be hooked for the button. On click, the functions filters the array for elements that fit the criteria set by the user input. As well a state variable was assigned for the output. (It's hard to test the function since i need to wait for the data to load otherwise it returns an empty array.)